requirements.txt
requests
fusionauth-client
python-dotenv
pika
